+++ src/test/java/net/revelc/code/formatter/xml/lib/FormatterTest.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package net.revelc.code.formatter.xml;
+package net.revelc.code.formatter.xml.lib;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
